Correct computation of H/W ratio in Bora (Igor Zivanovic).
* Bug: In Bora/SlicingDataModel, the division was done with integers,
    leading so there was a rounding *before* we casted to double.
      Now, use the BoxSet::getRatio() method and cast to double *before*
    dividing.

+++ bora/src/bora/BoxSet.h
@@ -87,7 +87,7 @@ namespace Bora {
   inline DbU::Unit    BoxSet::getWidth          () const { return _width; }
   inline unsigned int BoxSet::getCpt            () const { return _cpt; }
   inline void         BoxSet::incrementCpt      ()       { _cpt++; }
-  inline double       BoxSet::getRatio          ()       { return _width/_height; }
+  inline double       BoxSet::getRatio          ()       { return (double)_width/(double)_height; }
   inline double       BoxSet::getArea           ()       { return _width*_height; }
   inline unsigned int BoxSet::getType           () const { return UnknownType; }
   inline double       BoxSet::getOccupationArea () const { return 100; }
